Clear right of prompt #6
Clears any text to the right of the prompt. This means that when the prompt is updated (by using the fish directory jumping, for example) if the new prompt is shorter than the last, you don't have anything left over on the right-hand side. Resolves #6

Uses the `cleareol EL0` VT100 escape code for doing this and should be compatible with any VT100 compatible terminal emulator. This has been tested with `urxvt` and `xterm`.